+++ CHANGELOG.md
@@ -17,6 +17,8 @@ All user visible changes to this project will be documented in this file. This p
     - Always display online status on desktop. ([#702], [#681])
     - Style page:
         - Redesigned widgets tab. ([#695], [#632])
+    - Chat page:
+        - Display read partially message status in groups. ([#703], [#666])
 
 ### Fixed
 
@@ -24,9 +26,11 @@ All user visible changes to this project will be documented in this file. This p
     - Invalid caller name in popup calls. ([#711])
 
 [#632]: /../../issues/632
+[#666]: /../../issues/666
 [#681]: /../../issues/681
 [#695]: /../../pull/695
 [#702]: /../../pull/702
+[#703]: /../../pull/703
 [#711]: /../../pull/711
 
 
@@ -46,7 +50,6 @@ All user visible changes to this project will be documented in this file. This p
         - Display message field while loading. ([#662], [#634])
         - Display small images smaller. ([#688], [#653])
         - Message attachments and replies editing. ([#671], [#557])
-        - Display single mark as blue, if not the whole Chat-group read message. ([#703], [#666])
     - Disabled larger fonts accessibility setting temporary. ([#679])
     - Home page:
         - Redesigned introduction modal. ([#668], [#633])
@@ -68,13 +71,11 @@ All user visible changes to this project will be documented in this file. This p
 [#653]: /../../issues/653
 [#662]: /../../pull/662
 [#663]: /../../pull/663
-[#666]: /../../issues/666
 [#668]: /../../pull/668
 [#671]: /../../pull/671
 [#672]: /../../pull/672
 [#679]: /../../pull/679
 [#688]: /../../pull/688
-[#703]: /../../pull/703